refactory code.

+++ src/EQueue/Remoting/IRemotingClient.cs
@@ -7,9 +7,9 @@ namespace EQueue.Remoting
     {
         void Start();
         void Shutdown();
-        RemotingResponse InvokeSync(string address, RemotingRequest request, long timeoutMillis);
-        Task<RemotingResponse> InvokeAsync(string address, RemotingRequest request, long timeoutMillis);
-        void InvokeOneway(string address, RemotingRequest request, long timeoutMillis);
+        RemotingResponse InvokeSync(string address, RemotingRequest request, int timeoutMillis);
+        Task<RemotingResponse> InvokeAsync(string address, RemotingRequest request, int timeoutMillis);
+        void InvokeOneway(string address, RemotingRequest request, int timeoutMillis);
         void RegisterRequestProcessor(int requestCode, IRequestProcessor requestProcessor);
     }
 }
